Move PrintDurationSpanExporter to new file
