+++ playground/streaming/synthesizer/synthesize.py
@@ -1,12 +1,10 @@
 import time
 import argparse
 from typing import Optional
-from collections import defaultdict
 from vocode.streaming.agent.bot_sentiment_analyser import BotSentiment
 from vocode.streaming.models.message import BaseMessage
 from vocode.streaming.models.synthesizer import (
     AzureSynthesizerConfig,
-    ElevenLabsSynthesizerConfig,
     GoogleSynthesizerConfig,
     PlayHtSynthesizerConfig,
     RimeSynthesizerConfig,
@@ -31,23 +29,9 @@ if __name__ == "__main__":
     if args.trace:
         from opentelemetry import trace
         from opentelemetry.sdk.trace import TracerProvider
-        from opentelemetry.sdk.trace.export import SimpleSpanProcessor, SpanExporter
+        from opentelemetry.sdk.trace.export import SimpleSpanProcessor
         from opentelemetry.sdk.resources import Resource
-
-        class PrintDurationSpanExporter(SpanExporter):
-            def __init__(self):
-                super().__init__()
-                self.spans = defaultdict(list)
-
-            def export(self, spans):
-                for span in spans:
-                    duration_ns = span.end_time - span.start_time
-                    duration_s = duration_ns / 1e9
-                    self.spans[span.name].append(duration_s)
-
-            def shutdown(self):
-                for name, durations in self.spans.items():
-                    print(f"{name}: {sum(durations) / len(durations)}")
+        from playground.streaming.tracing_utils import PrintDurationSpanExporter
 
         trace.set_tracer_provider(TracerProvider(resource=Resource.create({})))
         span_exporter = PrintDurationSpanExporter()
